changes on genenv to ui/src/app/applications/components/application-summary/application-summary.tsx,ui/src/app/shared/components/array-input/array-input-field.tsx,ui/src/app/shared/components/array-input/array-input.tsx,
